requirements.txt
fastapi
uvicorn
joblib
numpy
pandas
scikit-learn
xgboost
optuna
pydantic
pyyaml
